+++ src/js/MovieApiSevice.js
@@ -23,7 +23,7 @@ const BASE_URL = 'https://api.themoviedb.org/3/';
 // }
 
 export async function downloadGenresIdList() {
-  const { data } = await axios(`${BASE_URL}/genre/movie/list?api_key=${API_KEY}`);
+  const { data } = await axios(`${BASE_URL}genre/movie/list?api_key=${API_KEY}`);
   return data.genres;
 }
 
@@ -35,10 +35,18 @@ export async function fetchSearch(additionalUrl, page) {
 
 export async function querySearch(query, page) {
   const searchMoviePath = 'search/movie';
-  // return await axios(
-  //   `${BASE_URL}${searchMoviePath}?query=${query}&api_key=${API_KEY}&page=${page}`);
   const { data } = await axios(
     `${BASE_URL}${searchMoviePath}?query=${query}&api_key=${API_KEY}&page=${page}`
   );
   return data;
 }
+
+// export async function fetchSearch(additionalUrl, page, queryKey = '') {
+//   if (queryKey) {
+//     queryKey = '&query=' + queryKey;
+//   }
+//   const { data } = await axios(
+//     `${BASE_URL}${additionalUrl}?api_key=${API_KEY}&page=${page}${queryKey}`
+//   );
+//   return data;
+// }
