+++ src/sass/components/_header.scss
@@ -131,11 +131,19 @@
   font-size: 14px;
   line-height: 1.14;
   text-align: center;
-  @include tablet{
+  @include tablet {
     display: flex;
     justify-content: center;
   }
 }
 .footer_link {
   color: #545454;
+}
+.error_message {
+  position: absolute;
+  top: 20px;
+  color: #ff001b;
+  font-size: 14px;
+  line-height: 16px;
+  text-align: center;
 }
